requirements.txt
# Flask and extensions
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-SocketIO==5.3.6

# Database
SQLAlchemy==2.0.42
psycopg2-binary
alembic==1.16.4

# Web Server (for production)
gunicorn
eventlet

# Data processing and file handling
pandas==2.3.1
openpyxl==3.1.5
Pillow==11.3.0
et-xmlfile==2.0.0
xlrd
python-docx

# QR Code generation
qrcode==8.2

# Utilities
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
blinker==1.9.0
click==8.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
numpy==2.3.2
packaging==25.0
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2

# Testing
pytest==8.4.1
iniconfig==2.1.0
pluggy==1.6.0

# Dependencies for other packages
colorama==0.4.6
greenlet==3.2.3
Pygments==2.19.2
setuptools>=78.1.1
pip-audit